requirements.txt
asgiref==3.3.1
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.3.0
autopep8==1.5.3
certifi==2020.12.5
colorama==0.4.3
Django==3.0.3
django-filter==2.3.0
django-rest-auth==0.9.5
django-rest-passwordreset==1.1.0
djangorestframework==3.11.1
Faker==4.17.1
importlib-metadata==3.3.0
iniconfig==1.1.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.6.0
packaging==20.8
pluggy==0.13.1
py==1.10.0
pycodestyle==2.6.0
pylint==2.5.3
pyparsing==2.4.7
pytest==6.0.1
pytest-django==3.9.0
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
wincertstore==0.2
wrapt==1.12.1
zipp==3.4.0
